lab changes + bev

+++ src/Controller/CreationHeaderController.java
@@ -6,12 +6,16 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.fxml.Initializable;
 import javafx.scene.Parent;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.layout.Pane;
+import javafx.scene.layout.Region;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
+import java.util.Optional;
 import java.util.ResourceBundle;
 
 public class CreationHeaderController implements Initializable {
@@ -47,13 +51,26 @@ public class CreationHeaderController implements Initializable {
 
     //Action for clicking "Delete" button, launches pop up to confirm before deleting the creation
     public void deleteCreation() {
-        model.deleteDirectory();
+    	Alert pbAlert = new Alert(Alert.AlertType.CONFIRMATION);
+        pbAlert.setTitle("Are you sure?");
+        pbAlert.setHeaderText("Are you sure you want to delete \"" + model.getCurrentItem() + "\" ?");
+        pbAlert.setContentText("Press \"Yes\" to delete, or \"No\" to cancel");
+        pbAlert.getDialogPane().setMinHeight(Region.USE_PREF_SIZE);
+        ButtonType yes = new ButtonType("Yes");
+        ButtonType no = new ButtonType("No");
+        pbAlert.getButtonTypes().setAll(yes, no);
+
+        Optional<ButtonType> result = pbAlert.showAndWait();
+        if (result.get() == yes) {
+        	model.deleteDirectory();
+        }
+        
     }
 
     //Helper method for setting the pane, takes the desired pane and a boolean for disabling buttons as arguments
     private void setPane(String name, Boolean bool) {
         try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource(".." + _fileSeperator + "View" + _fileSeperator + name + ".fxml"));
+            FXMLLoader loader = new FXMLLoader(getClass().getResource(_fileSeperator + "View" + _fileSeperator + name + ".fxml"));
             Parent root = loader.load(); //needs to have a different controller because initialise is causing an infinite loop.
             Main.mainPane.getChildren().setAll(root);
             /*
